requirements.txt
google-genai==1.18.0
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1